Move normalizeMeta to MetaUtils.{cpp,h}, use in TransactionMetaFrame
