+++ src/main/dumpxdr.cpp
@@ -9,6 +9,7 @@
 #include "util/Decoder.h"
 #include "util/Fs.h"
 #include "util/GlobalChecks.h"
+#include "util/MetaUtils.h"
 #include "util/XDRCereal.h"
 #include "util/XDROperators.h"
 #include "util/XDRStream.h"
@@ -165,86 +166,6 @@ printOneXdr(xdr::opaque_vec<> const& o, std::string const& desc, bool compact)
     T tmp;
     xdr::xdr_from_opaque(o, tmp);
     std::cout << xdr_to_string(tmp, desc, compact) << std::endl;
-}
-
-struct CmpLedgerEntryChanges
-{
-    int
-    remap(LedgerEntryChangeType let)
-    {
-        // order that we want is:
-        // LEDGER_ENTRY_STATE, LEDGER_ENTRY_CREATED,
-        // LEDGER_ENTRY_UPDATED, LEDGER_ENTRY_REMOVED
-        static constexpr std::array<int, 4> reindex = {1, 2, 3, 0};
-        return reindex[let];
-    }
-
-    LedgerKey
-    getKeyFromChange(LedgerEntryChange const& change)
-    {
-        LedgerKey res;
-        switch (change.type())
-        {
-        case LEDGER_ENTRY_STATE:
-            res = LedgerEntryKey(change.state());
-            break;
-        case LEDGER_ENTRY_CREATED:
-            res = LedgerEntryKey(change.created());
-            break;
-        case LEDGER_ENTRY_UPDATED:
-            res = LedgerEntryKey(change.updated());
-            break;
-        case LEDGER_ENTRY_REMOVED:
-            res = change.removed();
-            break;
-        }
-        return res;
-    }
-
-    bool
-    operator()(LedgerEntryChange const& l, LedgerEntryChange const& r)
-    {
-        auto lT =
-            std::make_tuple(getKeyFromChange(l), remap(l.type()), xdrSha256(l));
-        auto rT =
-            std::make_tuple(getKeyFromChange(r), remap(r.type()), xdrSha256(r));
-        return lT < rT;
-    }
-};
-
-void
-sortChanges(LedgerEntryChanges& c)
-{
-    std::sort(c.begin(), c.end(), CmpLedgerEntryChanges());
-}
-
-void
-normalizeOps(xdr::xvector<OperationMeta>& oms)
-{
-    for (auto& om : oms)
-    {
-        sortChanges(om.changes);
-    }
-}
-
-void
-normalizeMeta(TransactionMeta& m)
-{
-    switch (m.v())
-    {
-    case 0:
-        normalizeOps(m.operations());
-        break;
-    case 1:
-        sortChanges(m.v1().txChanges);
-        normalizeOps(m.v1().operations);
-        break;
-    case 2:
-        sortChanges(m.v2().txChangesBefore);
-        sortChanges(m.v2().txChangesAfter);
-        normalizeOps(m.v2().operations);
-        break;
-    }
 }
 
 void
@@ -481,8 +402,8 @@ signtxn(std::string const& filename, std::string netId, bool base64)
             xdr::xdr_from_opaque(d, txenv);
             auto& signatures = txbridge::getSignatures(txenv);
             if (signatures.size() == signatures.max_size())
-                throw std::runtime_error(
-                    "Envelope already contains maximum number of signatures");
+                throw std::runtime_error("Envelope already contains "
+                                         "maximum number of signatures");
 
             SecretKey sk(SecretKey::fromStrKeySeed(readSecret(
                 fmt::format(FMT_STRING("Secret key seed [network id: '{}']: "),
@@ -494,9 +415,9 @@ signtxn(std::string const& filename, std::string netId, bool base64)
             {
             case ENVELOPE_TYPE_TX_V0:
                 payload.taggedTransaction.type(ENVELOPE_TYPE_TX);
-                // TransactionV0 and Transaction always have the same signatures
-                // so there is no reason to check versions here, just always
-                // convert to Transaction
+                // TransactionV0 and Transaction always have the same
+                // signatures so there is no reason to check versions here,
+                // just always convert to Transaction
                 payload.taggedTransaction.tx() =
                     txbridge::convertForV13(txenv).v1().tx;
                 break;
